feat: banner 부분 ui 컴포넌트 분리 및 JSDocs 작성 [SN-42]
